requirements.txt
sqlalchemy==1.3.15
psycopg2-binary==2.9.3
grpcio==1.59.0
grpcio-tools==1.59.0
python-dotenv==0.14.0
efipay==1.0.0